Add static cast for size_t to uint64_t: #0000039

+++ knp/backends/cpu/cpu-library/include/knp/backends/cpu-library/impl/delta_synapse_projection_impl.h
@@ -135,7 +135,7 @@ void calculate_projection_part_impl(
     const std::unordered_map<knp::core::Step, size_t> &message_in_data, MessageQueue &future_messages, uint64_t step_n,
     uint64_t part_start, uint64_t part_size, std::mutex &mutex)
 {
-    size_t part_end = std::min(part_start + part_size, projection.size());
+    size_t part_end = std::min(part_start + part_size, static_cast<uint64_t>(projection.size()));
     std::vector<std::pair<uint64_t, knp::core::messaging::SynapticImpact>> container;
     WeightUpdateStdpMp<DeltaLikeSynapse>::init_projection_part(projection, message_in_data, step_n);
     for (size_t synapse_index = part_start; synapse_index < part_end; ++synapse_index)
